Add resume wrapper review regressions (#9200)
