+++ cmuxTests/AgentResumeReturnShellStartupTests.swift
@@ -238,88 +238,6 @@ struct AgentResumeReturnShellStartupTests {
         }
     }
 
-    @Test("launcher cwd and cwd-sensitive resume argv use the same restored directory")
-    func launcherAndResumeCommandShareRestoredWorkingDirectory() throws {
-        let fileManager = FileManager.default
-        let root = fileManager.temporaryDirectory
-            .appendingPathComponent("cmux-9200-retarget-\(UUID().uuidString)", isDirectory: true)
-        let savedDirectory = root.appendingPathComponent("saved", isDirectory: true)
-        let restoredDirectory = root.appendingPathComponent("restored", isDirectory: true)
-        try fileManager.createDirectory(at: restoredDirectory, withIntermediateDirectories: true)
-        defer { try? fileManager.removeItem(at: root) }
-        let registration = CmuxVaultAgentRegistration(
-            id: "cwd-agent",
-            name: "CWD Agent",
-            detect: CmuxVaultAgentDetectRule(processName: "cwd-agent"),
-            sessionIdSource: .argvOption("--session"),
-            resumeCommand: "{{executable}} --cwd {{cwd}} --session {{sessionId}}",
-            cwd: .preserve
-        )
-        let snapshot = SessionRestorableAgentSnapshot(
-            kind: .custom("cwd-agent"),
-            sessionId: "session-9200",
-            workingDirectory: savedDirectory.path,
-            launchCommand: AgentLaunchCommandSnapshot(
-                launcher: "cwd-agent",
-                executablePath: "cwd-agent",
-                arguments: ["cwd-agent"],
-                workingDirectory: savedDirectory.path,
-                environment: nil,
-                capturedAt: 123,
-                source: "test"
-            ),
-            registration: registration
-        )
-
-        let input = try #require(snapshot.resumeStartupInput(
-            fileManager: fileManager,
-            temporaryDirectory: root,
-            restoringWorkingDirectory: restoredDirectory.path
-        ))
-        let words = TerminalStartupWorkingDirectoryPrefix.shellWordRanges(
-            input.trimmingCharacters(in: .whitespacesAndNewlines)
-        ).map(\.value)
-        let launcherPath = try #require(words.last)
-        let script = try String(contentsOfFile: launcherPath, encoding: .utf8)
-        #expect(script.contains("cd -- \(TerminalStartupShellQuoting.singleQuoted(restoredDirectory.path))"))
-        #expect(script.contains("'--cwd' '\(restoredDirectory.path)'"))
-        #expect(!script.contains(savedDirectory.path))
-    }
-
-    @Test("restored logical cwd survives physical-path shell reports")
-    @MainActor
-    func restoredLogicalWorkingDirectorySurvivesResolvedReports() throws {
-        let fileManager = FileManager.default
-        let root = fileManager.temporaryDirectory
-            .appendingPathComponent("cmux-9200-logical-cwd-\(UUID().uuidString)", isDirectory: true)
-        let physicalDirectory = root.appendingPathComponent("physical", isDirectory: true)
-        let logicalDirectory = root.appendingPathComponent("logical", isDirectory: true)
-        try fileManager.createDirectory(at: physicalDirectory, withIntermediateDirectories: true)
-        try fileManager.createSymbolicLink(at: logicalDirectory, withDestinationURL: physicalDirectory)
-        defer { try? fileManager.removeItem(at: root) }
-
-        let source = Workspace()
-        let sourcePanelId = try #require(source.focusedPanelId)
-        let bindingIndex = SurfaceResumeBindingIndex(bindingsByPanel: [
-            .init(workspaceId: source.id, panelId: sourcePanelId): SurfaceResumeBindingSnapshot(
-                command: ":",
-                cwd: logicalDirectory.path,
-                source: "process-detected",
-                autoResume: true
-            ),
-        ])
-        let restored = Workspace()
-        restored.restoreSessionSnapshot(source.sessionSnapshot(
-            includeScrollback: false,
-            surfaceResumeBindingIndex: bindingIndex
-        ))
-        let restoredPanelId = try #require(restored.focusedPanelId)
-
-        #expect(restored.updatePanelDirectory(panelId: restoredPanelId, directory: physicalDirectory.path))
-        #expect(restored.updatePanelDirectory(panelId: restoredPanelId, directory: physicalDirectory.path))
-        #expect(restored.currentDirectory == logicalDirectory.path)
-    }
-
     @Test("auto-resume returns to the normally initialized login shell")
     func autoResumeReturnsToNormallyInitializedLoginShell() throws {
         let fileManager = FileManager.default
