+++ cmuxTests/ResumeLauncherCwdConsistencyTests.swift
@@ -1,0 +1,174 @@
+import Foundation
+import Testing
+
+#if canImport(cmux_DEV)
+    @testable import cmux_DEV
+#elseif canImport(cmux)
+    @testable import cmux
+#endif
+
+@Suite("Resume launcher cwd consistency")
+struct ResumeLauncherCwdConsistencyTests {
+    @Test("launcher cwd and cwd-sensitive resume argv use the same restored directory")
+    func launcherAndResumeCommandShareRestoredWorkingDirectory() throws {
+        let fileManager = FileManager.default
+        let root = fileManager.temporaryDirectory
+            .appendingPathComponent("cmux-9200-retarget-\(UUID().uuidString)", isDirectory: true)
+        let savedDirectory = root.appendingPathComponent("saved", isDirectory: true)
+        let restoredDirectory = root.appendingPathComponent("restored", isDirectory: true)
+        try fileManager.createDirectory(at: restoredDirectory, withIntermediateDirectories: true)
+        defer { try? fileManager.removeItem(at: root) }
+        let registration = CmuxVaultAgentRegistration(
+            id: "cwd-agent",
+            name: "CWD Agent",
+            detect: CmuxVaultAgentDetectRule(processName: "cwd-agent"),
+            sessionIdSource: .argvOption("--session"),
+            resumeCommand: "{{executable}} --cwd {{cwd}} --session {{sessionId}}",
+            cwd: .preserve
+        )
+        let snapshot = SessionRestorableAgentSnapshot(
+            kind: .custom("cwd-agent"),
+            sessionId: "session-9200",
+            workingDirectory: savedDirectory.path,
+            launchCommand: AgentLaunchCommandSnapshot(
+                launcher: "cwd-agent",
+                executablePath: "cwd-agent",
+                arguments: ["cwd-agent"],
+                workingDirectory: savedDirectory.path,
+                environment: nil,
+                capturedAt: 123,
+                source: "test"
+            ),
+            registration: registration
+        )
+
+        let input = try #require(snapshot.resumeStartupInput(
+            fileManager: fileManager,
+            temporaryDirectory: root,
+            restoringWorkingDirectory: restoredDirectory.path
+        ))
+        let words = TerminalStartupWorkingDirectoryPrefix.shellWordRanges(
+            input.trimmingCharacters(in: .whitespacesAndNewlines)
+        ).map(\.value)
+        let launcherPath = try #require(words.last)
+        let script = try String(contentsOfFile: launcherPath, encoding: .utf8)
+        #expect(script.contains("cd -- \(TerminalStartupShellQuoting.singleQuoted(restoredDirectory.path))"))
+        #expect(script.contains("'--cwd' '\(restoredDirectory.path)'"))
+        #expect(!script.contains(savedDirectory.path))
+    }
+
+    @Test("restored logical cwd survives physical-path shell reports")
+    @MainActor
+    func restoredLogicalWorkingDirectorySurvivesResolvedReports() throws {
+        let fileManager = FileManager.default
+        let root = fileManager.temporaryDirectory
+            .appendingPathComponent("cmux-9200-logical-cwd-\(UUID().uuidString)", isDirectory: true)
+        let physicalDirectory = root.appendingPathComponent("physical", isDirectory: true)
+        let logicalDirectory = root.appendingPathComponent("logical", isDirectory: true)
+        try fileManager.createDirectory(at: physicalDirectory, withIntermediateDirectories: true)
+        try fileManager.createSymbolicLink(at: logicalDirectory, withDestinationURL: physicalDirectory)
+        defer { try? fileManager.removeItem(at: root) }
+
+        let source = Workspace()
+        let sourcePanelId = try #require(source.focusedPanelId)
+        let bindingIndex = SurfaceResumeBindingIndex(bindingsByPanel: [
+            .init(workspaceId: source.id, panelId: sourcePanelId): SurfaceResumeBindingSnapshot(
+                command: ":",
+                cwd: logicalDirectory.path,
+                source: "process-detected",
+                autoResume: true
+            ),
+        ])
+        let restored = Workspace()
+        restored.restoreSessionSnapshot(source.sessionSnapshot(
+            includeScrollback: false,
+            surfaceResumeBindingIndex: bindingIndex
+        ))
+        let restoredPanelId = try #require(restored.focusedPanelId)
+
+        #expect(restored.updatePanelDirectory(panelId: restoredPanelId, directory: physicalDirectory.path))
+        #expect(restored.updatePanelDirectory(panelId: restoredPanelId, directory: physicalDirectory.path))
+        #expect(restored.currentDirectory == logicalDirectory.path)
+    }
+
+    @Test("remote inline resume strips the saved cwd after retargeting")
+    func remoteInlineResumeRemovesStaleSavedWorkingDirectoryOption() throws {
+        let savedDirectory = "/home/dev/repo"
+        let restoredDirectory = "/home/dev/repo/sub"
+        let snapshot = SessionRestorableAgentSnapshot(
+            kind: .codex,
+            sessionId: "remote-session-9200",
+            workingDirectory: savedDirectory,
+            launchCommand: AgentLaunchCommandSnapshot(
+                launcher: "codex",
+                executablePath: "codex",
+                arguments: ["codex", "--cd", savedDirectory],
+                workingDirectory: savedDirectory,
+                environment: nil,
+                capturedAt: 123,
+                source: "remote"
+            )
+        )
+
+        let input = try #require(snapshot.resumeStartupInput(
+            allowLauncherScript: false,
+            allowOversizedInlineInput: true,
+            restoringWorkingDirectory: restoredDirectory
+        ))
+        #expect(input.contains("cd -- '\(restoredDirectory)'"))
+        #expect(!input.contains(savedDirectory))
+    }
+
+    @Test("restored terminal command wrapper sources login files once")
+    func restoredTerminalCommandWrapperAvoidsNestedLoginStartup() throws {
+        let fileManager = FileManager.default
+        let root = fileManager.temporaryDirectory
+            .appendingPathComponent("cmux-9200-tmux-shell-\(UUID().uuidString)", isDirectory: true)
+        let home = root.appendingPathComponent("home", isDirectory: true)
+        let output = root.appendingPathComponent("counts.txt", isDirectory: false)
+        try fileManager.createDirectory(at: home, withIntermediateDirectories: true)
+        defer { try? fileManager.removeItem(at: root) }
+        try """
+        export CMUX_9200_ZPROFILE_COUNT=$(( ${CMUX_9200_ZPROFILE_COUNT:-0} + 1 ))
+
+        """.write(to: home.appendingPathComponent(".zprofile"), atomically: true, encoding: .utf8)
+        try """
+        export CMUX_9200_ZLOGIN_COUNT=$(( ${CMUX_9200_ZLOGIN_COUNT:-0} + 1 ))
+
+        """.write(to: home.appendingPathComponent(".zlogin"), atomically: true, encoding: .utf8)
+        let command = """
+        print -r -- "profile=${CMUX_9200_ZPROFILE_COUNT:-0} login=${CMUX_9200_ZLOGIN_COUNT:-0}" > \(TerminalStartupShellQuoting.singleQuoted(output.path))
+        """
+        let startupCommand = try #require(OneShotTerminalLauncherStore(
+            fileManager: fileManager,
+            temporaryDirectory: root
+        ).writeStartupCommand(command: command, workingDirectory: nil))
+        let words = TerminalStartupWorkingDirectoryPrefix.shellWordRanges(startupCommand).map(\.value)
+        #expect(Array(words.prefix(3)) == ["/usr/bin/env", "/bin/zsh", "-f"])
+
+        let process = Process()
+        process.executableURL = URL(fileURLWithPath: "/bin/bash")
+        process.arguments = ["--noprofile", "--norc", "-c", "exec -l \(startupCommand)"]
+        process.currentDirectoryURL = root
+        process.environment = [
+            "HOME": home.path,
+            "LOGNAME": NSUserName(),
+            "PATH": "/usr/bin:/bin",
+            "SHELL": "/bin/zsh",
+            "USER": NSUserName(),
+            "ZDOTDIR": home.path,
+        ]
+        let error = Pipe()
+        process.standardOutput = FileHandle.nullDevice
+        process.standardError = error
+        try process.run()
+        process.waitUntilExit()
+        let stderr = String(
+            data: error.fileHandleForReading.readDataToEndOfFile(),
+            encoding: .utf8
+        ) ?? ""
+
+        #expect(process.terminationStatus == 0, Comment(rawValue: stderr))
+        #expect(try String(contentsOf: output, encoding: .utf8) == "profile=1 login=1\n")
+    }
+}
